Начала опримизацию домашней страницы используя hilights React DevTools
