+++ project/src/components/movie-page-top-block/movie-page-top-block.tsx
@@ -1,19 +1,22 @@
-import {Link} from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import HeaderLogo from '../header-logo/header-logo';
 import HeaderLogin from '../../components/header-login/header-login';
-import FilmCardDescription from '../film-card-description/film-card-description';
 import {Film} from '../../types/films';
 import { useAppSelector } from '../../hooks';
 import { useEffect, useState } from 'react';
+import AddToMyListButton from '../add-to-my-list-button/add-to-my-list-button';
+import { AppRoute } from '../../const';
 
 type MoviePageTopBlockProps = {
     film: Film
   };
 
 function MoviePageTopBlock({film}: MoviePageTopBlockProps):JSX.Element {
+  const {promoFilm} = useAppSelector((state) => state);
   const [isAuth, setIsAuth] = useState<boolean>(false);
   const {authorizationStatus} = useAppSelector((state) => state);
   const {backgroundImage, name} = film;
+  const navigate = useNavigate();
 
   useEffect (() => {
     setIsAuth(authorizationStatus === 'AUTH');
@@ -33,9 +36,30 @@ function MoviePageTopBlock({film}: MoviePageTopBlockProps):JSX.Element {
       </header>
 
       <div className="film-card__wrap">
-        <FilmCardDescription film={film}>
-          {isAuth ? <Link to='review' className="btn film-card__button">Add review</Link>: ''}
-        </FilmCardDescription>
+        <div className="film-card__desc">
+          <h2 className="film-card__title">{promoFilm.name}</h2>
+          <p className="film-card__meta">
+            <span className="film-card__genre">{promoFilm.genre}</span>
+            <span className="film-card__year">{promoFilm.released}</span>
+          </p>
+
+          <div className="film-card__buttons">
+            {film.id}
+            <button onClick={() => {
+              navigate(`${AppRoute.Player}/${film.id}`);
+            }}
+            className="btn btn--play film-card__button"
+            type="button"
+            >
+              <svg viewBox="0 0 19 19" width="19" height="19">
+                <use xlinkHref="#play-s"></use>
+              </svg>
+              <span>Play</span>
+            </button>
+            <AddToMyListButton />
+            {isAuth ? <Link to={AppRoute.FilmReview} className="btn film-card__button">Add review</Link>: ''}
+          </div>
+        </div>
       </div>
     </div>
   );
